CurationEnvironment

Documentation for CurationEnvironment.

CurationEnvironment.Auction

CurationEnvironment.CRSPE

CurationEnvironment.CurationModel

CurationEnvironment.Curator

CurationEnvironment.DoubleCRSPE

CurationEnvironment.GraphEntity

CurationEnvironment.MinMaxCurator

CurationEnvironment.Subgraph

CurationEnvironment.auction

CurationEnvironment.best_response

CurationEnvironment.best_response

CurationEnvironment.best_response

CurationEnvironment.best_response

CurationEnvironment.best_response

CurationEnvironment.best_response

CurationEnvironment.bopt

CurationEnvironment.burntokens

CurationEnvironment.burntokens

CurationEnvironment.curate

CurationEnvironment.equity_proportion

CurationEnvironment.equity_proportion

CurationEnvironment.id

CurationEnvironment.latefees

CurationEnvironment.minttokens

CurationEnvironment.multiple_bidders

CurationEnvironment.payment

CurationEnvironment.payment

CurationEnvironment.popt

CurationEnvironment.shares

CurationEnvironment.shares

CurationEnvironment.single_bidder

CurationEnvironment.step

CurationEnvironment.step

CurationEnvironment.step

CurationEnvironment.step

CurationEnvironment.utility

CurationEnvironment.utility

CurationEnvironment.τ

CurationEnvironment.Auction — Type

Wrap a CurationModel in an auction.

CurationEnvironment.CRSPE — Type

CRSPE(m::Model)

A commit-reveal, second price auction has three stages:

Bid: Each participant privately bids (perhaps multiple times) via hash

commitment(s).

Reveal: Bidders (optionally) deposit tokens equal to their bid with a

hash verification.

Allocate: The highest bidder is reimbursed the difference between their

bid and the second highest bid and allocated shares in proportion to their payment. All others are reimbursed for their deposits.

CurationEnvironment.CurationModel — Type

Dynamics governing the interaction between a Subgraph and a Curator.

CurationEnvironment.Curator — Type

Curator <: AbstractCurator

Signal tokens on a Subgraph as per a private valuation.

Curator is an entity that signals tokens on a Subgraph to demonstrate the value of the subgraph to indexers. Curators are paid via query fees when on a subgraph. Curator id estimates the subgraph valuations as v̂s and owns ςs shares on each subgraph. The curator has σ stake to spend.

Constructors

Curator{M,S,T}(
    id::<:Integer,
    vs::NTuple{M,T},
    ςs::NTuple{M,T},
    σ::T
) where {S<:Integer, T<:Real}

Curator{M}(id::<:Integer, ̂vs::NTuple{M,<:Real}, ςs::NTuple{M,<:Real}, σ::<:Real)
Curator(id::<:Integer, ̂vs::NTuple{M,<:Real}, ςs::NTuple{M,<:Real}, σ::<:Real)
Curator{M}(id::<:Integer, ̂v::<:Real, ς::<:Real, σ::<:Real)

CurationEnvironment.DoubleCRSPE — Type

DoubleCRSPE(m::CRSPE)

The difference between DoubleCRSPE and CRSPE is that in DoubleCRSPE, we burn tokens as per a CRSPE auction as well.

This can be constructed by bypassing CRSPE and directly specifying the CurationModel. DoubleCRSPE(CommunitySignal()).

CurationEnvironment.GraphEntity — Type

A type of entity that interacts with The Graph protocol.

CurationEnvironment.MinMaxCurator — Type

Curator <: AbstractCurator

A curator that has both a min and max valuation.

MinMaxCurator is a Curator that has both a minimum valuation for each subgraph v̂mins and a maximum valuation per subgraph v̂maxs. Intuitively, v̂min for a subgraph is the minimum amount of signal you would want to see on a subgraph. v̂max is your true valuation of the subgraph. The other parameters are as given by Curator.

CurationEnvironment.Subgraph — Type

Subgraph <: GraphEntity

Blockchain data is organised into subgraphs.

Subgraph is an entity on which curators signal tokens. Subgraph id has signal v, shares ς and late fee parameter of τ. τ is related to the late fee rate by τ.

Constructors

Subgraph{S,T}(id::S, v::T, ς::T, τ::T) where {S<:Integer,T<:Real}
Subgraph(id::<:Integer, v::<:Real, ς::<:Real, τ::<:Real)

See also Curator

CurationEnvironment.auction — Method

auction(m::CRSPE, bids::Vector{<:Real}, cs::Vector{<:Curator}, s::Subgraph)

Runs a commit-reveal second-price auction to select which bid wins the right to curate on the subgraph. The curators cs make transactions t based on bids on the subgraph s.

CurationEnvironment.best_response — Method

best_response(m::CRSPE{CommunitySignal}, c::MinMaxCurator, s::Subgraph)

Find the best response for the min-max curator c on subgraph s.

CurationEnvironment.best_response — Method

best_response(::CRSPE{CommunitySignal}, v::Real, v̂::Real, τ::Real, ξ::Real)

Find the best response on the community signal model in a CRSPE auction for a subgraph with signal v and tax rate τ given the min-max curator believes the true value of the subgraph to be in the range v̂min and v̂max. The curator has the ratio ξ of the total shares on the subgraph and available stake σ.

CurationEnvironment.best_response — Method

best_response(m::CommunitySignal, c::Curator, s::Subgraph)

Find the best response for curator c on subgraph s.

CurationEnvironment.best_response — Method

best_response(m::CommunitySignal, c::MinMaxCurator, s::Subgraph)

Find the best response for the min-max curator c on subgraph s.

CurationEnvironment.best_response — Method

best_response(m::CommunitySignal, v::Real, v̂min::Real,
              v̂max::Real τ::Real, ξ::Real, σ::Real)

Find the best response on the community signal model for a subgraph with signal v and tax rate τ given the min-max curator believes the true value of the subgraph to be in the range v̂min and v̂max. The curator has the ratio ξ of the total shares on the subgraph and available stake σ.

CurationEnvironment.best_response — Method

best_response(m::CommunitySignal, v::Real, v̂::Real, τ::Real, ξ::Real, σ::Real)

Find the best response on the community signal model for a subgraph with signal v and tax rate τ given the curator believes the true value of the subgraph to be v̂. The curator has the ratio ξ of the total shares on the subgraph and available stake σ.

CurationEnvironment.bopt — Method

The optimal value to burn.

CurationEnvironment.burntokens — Method

burntokens(m::CRSPE, ps::Vector{<:Real}, cs::Vector{<:Curator}, s::Subgraph)

Curators cs who have negative payment in ps will burn tokens on the subgraph s.

CurationEnvironment.burntokens — Method

burntokens(m::DoubleCRSPE, ps::Vector{<:Real}, cs::Vector{<:Curator}, s::Subgraph)

Curators cs who have negative payment in ps will burn tokens on the subgraph s. In a DoubleCRSPE auction, they will burn tokens in a CRSPE auction.

CurationEnvironment.curate — Method

curate(m::CurationModel, p::Real, c::Curator, s::Subgraph)

A curator c curates tokens p on subgraph s as per model m.

See also Subgraph, Curator, CurationModel

CurationEnvironment.equity_proportion — Method

equity_proportion(::CommunitySignal, p::Real, v::Real, τ::Real)

The proportion of equity on a subgraph with signal v the curator will receive by paying amount p with tax rate τ.

CurationEnvironment.equity_proportion — Method

equity_proportion(model::CommunitySignal, p::Real, s::Subgraph)

The proportion of equity on a subgraph s the curator will receive by paying amount p.

CurationEnvironment.id — Method

The id of the GraphEntity.

CurationEnvironment.latefees — Method

latefees(m::CommunitySignal, p::Real, s::Subgraph)

The late fees a curator makes for paying p on subgraph s.

CurationEnvironment.minttokens — Method

minttokens(m::CRSPE, ps::Vector{<:Real}, cs::Vector{<:Curator}, s::Subgraph)

Curators cs who have positive payment in ps will mint tokens on the subgraph s.

CurationEnvironment.multiple_bidders — Method

multiple_bidders(m::CRSPE, bids::Vector{<:Real}, cs::Vector{<:Curator}, s::Subgraph)

When there are multiple bidders, the bidder who is willing to pay the most pays the price of the second-highest bid. The curators cs make transaction t based on bids on the subgraph s.

CurationEnvironment.payment — Method

payment(::CommunitySignal, x::Real, v::Real, τ::Real)

The payment needed to capture or burn x proportion of the equity for a subgraph with valuation v and tax rate τ.

CurationEnvironment.payment — Method

payment(model::CommunitySignal, x::Real, s::Subgraph)

The payment needed to capture x proportion of the equity for a subgraph s.

CurationEnvironment.popt — Method

The optimal value to mint.

CurationEnvironment.shares — Method

shares(model::CommunitySignal, x::Real, s::Real, v::Real, τ::Real)

How many new shares were minted when a curator executes a new transaction for equity proportion x on a subgraph with shares ς, signal v and tax rate τ.

CurationEnvironment.shares — Method

shares(model::CommunitySignal, x::Real, s::Subgraph)

How many new shares were minted when a curator executes a new transaction for equity proportion x on a subgraph s.

CurationEnvironment.single_bidder — Method

single_bidder(m::CRSPE, bids::Vector{<:Real}, cs::Vector{<:Curator}, s::Subgraph)

When there is only a single bidder, the bidder wins the auction. The curators cs make transaction t based on bids on the subgraph s.

CurationEnvironment.step — Method

step(m::CommunitySignal, π::Function, c::Tuple{AbstractCurator}, s::Subgraph)

Curators cs decide how much to curate on subgraph s by running the policy π.

Note that the order in which the curators execute is randomised each time step is called.

CurationEnvironment.step — Method

step(m::CRSPE, πs::Vector{<:Function}, cs::Vector{<:Curator}, s::Subgraph) → Tuple{Tuple{Curator}, Subgraph}

The curators cs bid in a CRSPE auction on subgraph s according to policies πs.

NOTE: Here, since we use a step function rather than async exec, the policies are executed together. This may be a bad assumption as burning can happen between auctions, whereas minting can only happen at discrete intervals with auctions.

CurationEnvironment.step — Method

step(m::DoubleCRSPE, πs::Vector{<:Function}, cs::Vector{<:Curator}, s::Subgraph) → Tuple{Tuple{Curator}, Subgraph}

The curators cs bid in a Double CRSPE auction on subgraph s according to policies πs.

NOTE: Here, since we use a step function rather than async exec, the policies are executed together. This may be a bad assumption as burning can happen between auctions, whereas minting can only happen at discrete intervals with auctions.

CurationEnvironment.step — Method

step(m::CommunitySignal, π::Function, c::AbstractCurator, s::Subgraph)

Curator c decides how much to curate on subgraph s by running the policy π.

CurationEnvironment.utility — Method

utility(m::CommunitySignal, p::Real, c::MinMaxCurator, s::Subgraph)

Utility on for payment p on subgraph s for the min-max curator c for model m.

CurationEnvironment.utility — Method

utility(m::CommunitySignal, p::Real, v̂min::Real, v̂max::Real, v::Real, τ::Real, ξ::Real)

Utility on for payment p for the min-max curator c for model m for a subgraph with signal v and tax rate τ given the min-max curator believes the true value of the subgraph to be in the range v̂min and v̂max. The curator has the ratio ξ of the total shares on the subgraph and available stake σ.

CurationEnvironment.τ — Method

Compute the fee parameter from the fee rate.