CurationEnvironment

Documentation for CurationEnvironment.

CurationEnvironment.AbstractCurator

CurationEnvironment.Auction

CurationEnvironment.Bid

CurationEnvironment.CRSPE

CurationEnvironment.CommunitySignal

CurationEnvironment.CurationModel

CurationEnvironment.Curator

CurationEnvironment.DoubleCRSPE

CurationEnvironment.GraphEntity

CurationEnvironment.MinMaxCurator

CurationEnvironment.Subgraph

CurationEnvironment.auction

CurationEnvironment.best_response

CurationEnvironment.best_response

CurationEnvironment.best_response

CurationEnvironment.burnshares

CurationEnvironment.burntokens

CurationEnvironment.curate

CurationEnvironment.equity_proportion

CurationEnvironment.equity_proportion

CurationEnvironment.latefees

CurationEnvironment.mintshares

CurationEnvironment.payment

CurationEnvironment.payment

CurationEnvironment.pmax

CurationEnvironment.popt

CurationEnvironment.shares

CurationEnvironment.shares

CurationEnvironment.step

CurationEnvironment.step

CurationEnvironment.step

CurationEnvironment.step

CurationEnvironment.utility

CurationEnvironment.utility

CurationEnvironment.winner

CurationEnvironment.AbstractCurator — Type

AbstractCurator is the abstract type for all curators.

All properties of an AbstractCurator will have an associated getter and setter. For example, the Curator concretion has methods: id, v̂s, ςs, and σ. If you pass in only the AbstractCurator, or for some fields, the AbstractCurator and an index i, then the method serves as a getter. If you pass in a value v in addition to the parameters of the getters, the method serves as a setter.

See also Curator

CurationEnvironment.Auction — Type

Wrap a CurationModel in an auction.

CurationEnvironment.Bid — Type

The bid in an Auction.

CurationEnvironment.CRSPE — Type

CRSPE <: Auction

A commit-reveal, second price auction has three stages:

Bid: Each participant privately bids (perhaps multiple times) via hash

commitment(s).

Reveal: Bidders (optionally) deposit tokens equal to their bid with a

hash verification.

Allocate: The highest bidder is reimbursed the difference between their

bid and the second highest bid and allocated shares in proportion to their payment. All others are reimbursed for their deposits.

Constructors

CRSPE{M<:Model}(m::M)
CRSPE(m::M)

CurationEnvironment.CommunitySignal — Type

CommunitySignal <: CurationModel

Reflect idea of of individuals contributing to a community asset.

Under this model, curators contribute tokens to the subgraph signal, which entitles them to shares. The shares entitle curators to rewards. Shares are created (minted) whenever a curator pays into the subgraph. Shares are destroyed (burned) whenever a curator withdraws funds from the subgraph.

Part of curation involves encouraging people to curate early, rather than to curate just before query fees come in. In the community signal model, we encourage early curation via a late fee, a small percentage which decreases the effectiveness of your current payment based on the amount already on the subgraph. The late fee amount is distributed to all curators who had entered before you, thus rewarding them for being early.

CurationEnvironment.CurationModel — Type

Dynamics governing the interaction between a Subgraph and a Curator.

CurationEnvironment.Curator — Type

Curator{M,S,T} <: AbstractCurator

Signal tokens on a Subgraph as per a private valuation.

Curator is an entity that signals tokens on a Subgraph to demonstrate the value of the subgraph to indexers. Curators are paid via query fees when on a subgraph. Curator id estimates the subgraph valuations as v̂s and owns ςs shares on each subgraph. The curator has σ stake to spend.

Constructors

Curator{M,S,T}(
    id::<:Integer,
    vs::NTuple{M,T},
    ςs::NTuple{M,T},
    σ::T
) where {S<:Integer, T<:Real}

Curator{M}(id::<:Integer, ̂vs::NTuple{M,<:Real}, ςs::NTuple{M,<:Real}, σ::<:Real)
Curator(id::<:Integer, ̂vs::NTuple{M,<:Real}, ςs::NTuple{M,<:Real}, σ::<:Real)
Curator{M}(id::<:Integer, ̂v::<:Real, ς::<:Real, σ::<:Real)

CurationEnvironment.DoubleCRSPE — Type

DoubleCRSPE(m::CRSPE)

The difference between DoubleCRSPE and CRSPE is that in DoubleCRSPE, we burn tokens as per a CRSPE auction as well.

This can be constructed by bypassing CRSPE and directly specifying the CurationModel. DoubleCRSPE(CommunitySignal()).

CurationEnvironment.GraphEntity — Type

A type of entity that interacts with The Graph protocol.

CurationEnvironment.MinMaxCurator — Type

MinMaxCurator{M,S,T} <: AbstractCurator

A curator that has both a min and max valuation.

MinMaxCurator is a Curator that has both a minimum valuation for each subgraph v̂mins and a maximum valuation per subgraph v̂maxs. Intuitively, v̂min for a subgraph is the minimum amount of signal you would want to see on a subgraph. v̂max is your true valuation of the subgraph. The other parameters are as given by Curator.

Constructors

MinMaxCurator(
    id::<:Integer,
    v̂mins::NTuple{M,<:Real},
    v̂maxs::NTuple{M,<:Real},
    ςs::NTuple{M,<:Real},
    σ::<:Real
) where {M}

MinMaxCurator{M,S,T}(
    id::S,
    v̂mins::NTuple{M,T},
    v̂maxs::NTuple{M,T},
    ςs::NTuple{M,T},
    σ::T
) where {M,S<:Integer,T<:Real}

CurationEnvironment.Subgraph — Type

Subgraph <: GraphEntity

Blockchain data is organised into subgraphs.

Subgraph is an entity on which curators signal tokens. Subgraph id has signal v, shares ς and late fee parameter of τ. τ is related to the late fee rate by τ.

Constructors

Subgraph{S,T}(id::S, v::T, ς::T, τ::T) where {S<:Integer,T<:Real}
Subgraph(id::<:Integer, v::<:Real, ς::<:Real, τ::<:Real)

See also Curator

CurationEnvironment.auction — Method

auction(m::CRSPE, bids::StructArray{<:CRSPEBid}, cs::Vector{<:MinMaxCurator}, s::Subgraph, t::Transaction)

Runs a commit-reveal second-price auction to select which bid wins the right to curate on the subgraph. The curators cs make transactions t based on bids on the subgraph s.

CurationEnvironment.best_response — Method

best_response(m::CRSPE{CommunitySignal}, c::MinMaxCurator, s::Subgraph)

Find the best response for the min-max curator c on subgraph s.

CurationEnvironment.best_response — Method

best_response(m::CommunitySignal, c::MinMaxCurator, s::Subgraph)

Find the best response for the min-max curator c on subgraph s.

CurationEnvironment.best_response — Method

best_response(m::CommunitySignal, v::Real, v̂min::Real,
              v̂max::Real τ::Real, ξ::Real, σ::Real)

Find the best response on the community signal model for a subgraph with signal v and tax rate τ given the min-max curator believes the true value of the subgraph to be in the range v̂min and v̂max. The curator has the ratio ξ of the total shares on the subgraph and available stake σ.

CurationEnvironment.burnshares — Method

burnshares(m::CRSPE, ps::StructArray{<:CRSPEBid}, cs::Vector{<:MinMaxCurator}, s::Subgraph)

Curators cs who have negative payment in ps will burn shares on the subgraph s.

CurationEnvironment.burntokens — Method

burntokens(m::DoubleCRSPE, ps::Vector{<:Real}, cs::Vector{<:Curator}, s::Subgraph)

Curators cs who have negative payment in ps will burn tokens on the subgraph s. In a DoubleCRSPE auction, they will burn tokens in a CRSPE auction.

CurationEnvironment.curate — Method

curate(m::Model, p::Real, c::Curator, s::Subgraph)

A curator c curates tokens p on subgraph s as per model m.

See also Subgraph, Curator, CurationModel

CurationEnvironment.equity_proportion — Method

equity_proportion(::CommunitySignal, p::Real, v::Real, τ::Real)

The proportion of equity on a subgraph with signal v the curator will receive by paying amount p with tax rate τ.

CurationEnvironment.equity_proportion — Method

equity_proportion(model::CommunitySignal, p::Real, s::Subgraph)

The proportion of equity on a subgraph s the curator will receive by paying amount p.

CurationEnvironment.latefees — Method

latefees(m::CommunitySignal, p::Real, s::Subgraph)

The late fees a curator makes for paying p on subgraph s.

CurationEnvironment.mintshares — Method

mintshares(m::CRSPE, ps::Vector{<:CRSPEBid}, cs::Vector{<:MinMaxCurator}, s::Subgraph)

Curators cs who have positive payment in ps will mint shares on the subgraph s.

CurationEnvironment.payment — Method

payment(::CommunitySignal, x::Real, v::Real, τ::Real)

The payment needed to capture or burn x proportion of the equity for a subgraph with valuation v and tax rate τ.

CurationEnvironment.payment — Method

payment(model::CommunitySignal, x::Real, s::Subgraph)

The payment needed to capture x proportion of the equity for a subgraph s.

CurationEnvironment.pmax — Method

pmax(::CommunitySignal, v::Real, v̂max::Real, τ::Real, ξ::Real, σ::Real)

The maximum profitable payment.

The subgraph has signal v and a fee rate τ. The curator has a max valuation of v̂max, stake σ and owns ξ proportion of shares on the subgraph.

CurationEnvironment.popt — Method

popt(::CommunitySignal, v::Real, v̂min::Real, v̂max::Real, τ::Real, ξ::Real)

The optimal amount to curate under the community signal model.

The subgraph has current signal v and fee rate τ. The curator has a min valuation v̂min, a max valuationv̂max, and stakeσ. The curator ownsξ` proportion of the shares on the subgraph.

See also utility, pmax.

CurationEnvironment.shares — Method

shares(model::CommunitySignal, x::Real, s::Real, v::Real, τ::Real)

How many new shares were minted when a curator executes a new transaction for equity proportion x on a subgraph with shares ς, signal v and tax rate τ.

CurationEnvironment.shares — Method

shares(model::CommunitySignal, x::Real, s::Subgraph)

How many new shares were minted when a curator executes a new transaction for equity proportion x on a subgraph s.

CurationEnvironment.step — Method

step(m::CommunitySignal, π::Function, c::Tuple{AbstractCurator}, s::Subgraph)

Curators cs decide how much to curate on subgraph s by running the policy π.

Note that the order in which the curators execute is randomised each time step is called.

CurationEnvironment.step — Method

step(m::CRSPE, πs::Vector{<:Function}, cs::Vector{<:AbstractCurator}, s::Subgraph) → Tuple{Tuple{Curator}, Subgraph}

The curators cs bid in a CRSPE auction on subgraph s according to policies πs.

NOTE: Here, since we use a step function rather than async exec, the policies are executed together. This may be a bad assumption as burning can happen between auctions, whereas minting can only happen at discrete intervals with auctions.

CurationEnvironment.step — Method

step(m::DoubleCRSPE, πs::Vector{<:Function}, cs::Vector{<:Curator}, s::Subgraph) → Tuple{Tuple{Curator}, Subgraph}

The curators cs bid in a Double CRSPE auction on subgraph s according to policies πs.

NOTE: Here, since we use a step function rather than async exec, the policies are executed together. This may be a bad assumption as burning can happen between auctions, whereas minting can only happen at discrete intervals with auctions.

CurationEnvironment.step — Method

step(m::CommunitySignal, π::Function, c::AbstractCurator, s::Subgraph)

Curator c decides how much to curate on subgraph s by running the policy π.

CurationEnvironment.utility — Method

utility(m::CommunitySignal, p::Real, c::MinMaxCurator, s::Subgraph)

Utility on for payment p on subgraph s for the min-max curator c for model m.

CurationEnvironment.utility — Method

utility(m::CommunitySignal, p::Real, v̂min::Real, v̂max::Real, v::Real, τ::Real, ξ::Real)

Utility on for payment p for the min-max curator c for model m for a subgraph with signal v and tax rate τ given the min-max curator believes the true value of the subgraph to be in the range v̂min and v̂max. The curator has the ratio ξ of the total shares on the subgraph and available stake σ.

CurationEnvironment.winner — Method

winner(m::CRSPE, bids::StructArray{<:CRSPEBid}, cs::Vector{<:MinMaxCurator}, s::Subgraph, t::Transaction)

Find the winner of the auction. The bidder who is willing to pay the most pays the price of the second-highest bid. The bidders cs make transaction t based on bids on the subgraph s.